fix(sse): extract Vertex error classifier and rebaseline frozen file sizes
